Formally save labels as ints, casting in bayes

+++ zoobot/estimators/bayesian_estimator_funcs.py
@@ -375,7 +375,7 @@ def get_eval_metric_ops(self, labels, predictions):
     tf.summary.histogram('total_votes', labels[1, :])
     assert labels.dtype == tf.float32
     assert predictions['prediction'].dtype == tf.float32
-    return {"rmse": tf.metrics.root_mean_squared_error(labels[0, :] / labels[1, :], predictions['prediction'])}
+    return {"rmse": tf.metrics.root_mean_squared_error(tf.cast(labels[0, :], dtype=tf.float32) / tf.cast(labels[1, :], dtype=tf.float32), predictions['prediction'])}
 
 def logging_hooks(model_config):
     train_tensors = {
